chore(auth): merge logout_redirect.go into handler.go
